# Perf: support sending JDWP commands asynchronously