--- a/com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/IMethodBreakpoint.java
+++ b/com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/IMethodBreakpoint.java
@@ -30,4 +30,11 @@ public interface IMethodBreakpoint extends IDebugResource {
     Object getProperty(Object key);
 
     void putProperty(Object key, Object value);
+
+    default void setAsync(boolean async) {
+    }
+
+    default boolean async() {
+        return false;
+    }
 }