--- a/com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/adapter/IDebugAdapterContext.java
+++ b/com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/adapter/IDebugAdapterContext.java
@@ -136,4 +136,10 @@ public interface IDebugAdapterContext {
     void setProcessId(long processId);
 
     long getProcessId();
+
+    void setThreadCache(ThreadCache cache);
+
+    ThreadCache getThreadCache();
+
+    boolean asyncJDWP();
 }