This PR provides some experimental performance tuning for remote debugging:
- Add a thread cache to reduce JDWP frequency for thread info. `threads` request is the most frequent DAP request, and adding a cache can save lots of unnecessary JDWP queries.
- Provide async mode to process JDWP commands in parallel. This hopes to improve the DAP responsiveness on high-latency networks.